requirements.txt
smartapi-python>=1.4.0
python-dotenv>=1.0.1
pyotp>=2.9.0
logzero>=1.7.0
pandas>=1.5.0
requests>=2.28.0
pyyaml>=6.0
pyarrow>=10.0.0
sqlalchemy>=1.4.0
pytest>=7.0.0
